Trait executor_types::ChunkExecutor

pub trait ChunkExecutor: Send + Sync {
    fn execute_chunk(
        &self, 
        txn_list_with_proof: TransactionListWithProof, 
        verified_target_li: LedgerInfoWithSignatures
    ) -> Result<(ProcessedVMOutput, Vec<TransactionToCommit>, Vec<ContractEvent>)>;

    fn apply_chunk(
        &self, 
        txn_output_list_with_proof: TransactionOutputListWithProof, 
        verified_target_li: LedgerInfoWithSignatures
    ) -> Result<(ProcessedVMOutput, Vec<TransactionToCommit>, Vec<ContractEvent>)>;

    fn commit_chunk(
        &self, 
        verified_target_li: LedgerInfoWithSignatures, 
        epoch_change_li: Option<LedgerInfoWithSignatures>, 
        output: ProcessedVMOutput, 
        txns_to_commit: Vec<TransactionToCommit>, 
        events: Vec<ContractEvent>
    ) -> Result<Vec<ContractEvent>>;

    fn execute_or_apply_chunk(
        &self, 
        first_version: u64, 
        transactions: Vec<Transaction>, 
        transaction_outputs: Option<Vec<TransactionOutput>>, 
        transaction_infos: Vec<TransactionInfo>
    ) -> Result<(ProcessedVMOutput, Vec<TransactionToCommit>, Vec<ContractEvent>)>;

    fn execute_and_commit_chunk(
        &self, 
        txn_list_with_proof: TransactionListWithProof, 
        verified_target_li: LedgerInfoWithSignatures, 
        epoch_change_li: Option<LedgerInfoWithSignatures>
    ) -> Result<Vec<ContractEvent>> { ... }

    fn apply_and_commit_chunk(
        &self, 
        txn_output_list_with_proof: TransactionOutputListWithProof, 
        verified_target_li: LedgerInfoWithSignatures, 
        epoch_change_li: Option<LedgerInfoWithSignatures>
    ) -> Result<Vec<ContractEvent>> { ... }
}

Required methods

fn execute_chunk(
&self,
txn_list_with_proof: TransactionListWithProof,
verified_target_li: LedgerInfoWithSignatures
) -> Result<(ProcessedVMOutput, Vec<TransactionToCommit>, Vec<ContractEvent>)>

Verifies the transactions based on the provided proofs and ledger info. If the transactions are valid, executes them and returns the executed result for commit.

fn apply_chunk(
&self,
txn_output_list_with_proof: TransactionOutputListWithProof,
verified_target_li: LedgerInfoWithSignatures
) -> Result<(ProcessedVMOutput, Vec<TransactionToCommit>, Vec<ContractEvent>)>

Similar to execute_chunk, but instead of executing transactions, apply the transaction outputs directly to get the executed result.

fn commit_chunk(
&self,
verified_target_li: LedgerInfoWithSignatures,
epoch_change_li: Option<LedgerInfoWithSignatures>,
output: ProcessedVMOutput,
txns_to_commit: Vec<TransactionToCommit>,
events: Vec<ContractEvent>
) -> Result<Vec<ContractEvent>>

Commit a previously executed chunks, returns a vector of reconfiguration events in the chunk.

fn execute_or_apply_chunk(
&self,
first_version: u64,
transactions: Vec<Transaction>,
transaction_outputs: Option<Vec<TransactionOutput>>,
transaction_infos: Vec<TransactionInfo>
) -> Result<(ProcessedVMOutput, Vec<TransactionToCommit>, Vec<ContractEvent>)>

Provided methods

fn execute_and_commit_chunk(
&self,
txn_list_with_proof: TransactionListWithProof,
verified_target_li: LedgerInfoWithSignatures,
epoch_change_li: Option<LedgerInfoWithSignatures>
) -> Result<Vec<ContractEvent>>

fn apply_and_commit_chunk(
&self,
txn_output_list_with_proof: TransactionOutputListWithProof,
verified_target_li: LedgerInfoWithSignatures,
epoch_change_li: Option<LedgerInfoWithSignatures>
) -> Result<Vec<ContractEvent>>