Crate executor_types

Structs

ExecutedTrees
A wrapper of the in-memory state sparse merkle tree and the transaction accumulator that represent a specific state collectively. Usually it is a state after executing a block.

ProcessedVMOutput
The output of Processing the vm output of a series of transactions to the parent in-memory state merkle tree and accumulator.

ProofReader

StateComputeResult
A structure that summarizes the result of the execution needed for consensus to agree on. The execution is responsible for generating the ID of the new state, which is returned in the result.

TransactionData
The entire set of data associated with a transaction. In addition to the output generated by VM which includes the write set and events, this also has the in-memory trees.

Enums

Error
Different reasons for proposal rejection

Traits

BlockExecutor

ChunkExecutor