Trait executor_types::ChunkExecutor

pub trait ChunkExecutor: Send {
    fn execute_and_commit_chunk(
        &mut self, 
        txn_list_with_proof: TransactionListWithProof, 
        verified_target_li: LedgerInfoWithSignatures, 
        epoch_change_li: Option<LedgerInfoWithSignatures>
    ) -> Result<Vec<ContractEvent>>;
}

Required methods

fn execute_and_commit_chunk(
&mut self,
txn_list_with_proof: TransactionListWithProof,
verified_target_li: LedgerInfoWithSignatures,
epoch_change_li: Option<LedgerInfoWithSignatures>
) -> Result<Vec<ContractEvent>>

Verifies the transactions based on the provided proofs and ledger info. If the transactions are valid, executes them and commits immediately if execution results match the proofs. Returns a vector of reconfiguration events in the chunk

Loading content...